Improved SRP and ease of function/file navigation
